requirements.txt
flask==2.2.5
tensorflow==2.10.0
scikit-learn==1.2.0
numpy==1.23.5
spacy==3.5.0
python-dotenv==0.21.0
faiss-cpu==1.7.3
sentence-transformers==2.2.2
pandas==1.5.3